Switch to a more modular project setting
We split up the spider into three distinct, orthogonal
modules. This way we can ensure division of tasks and
keep the code readable. Further we introduced sequelize
instead of the db moduel, which makes the db handling
easier and most importantly, better readable.
To enforce good readability and later on easy documentation
generation, eslint now enforces jsdoc comments and validates
them as well. Travis will be set up to generate the docs
in every build in a next iteration of the travis settings.

+++ deployment/setup.sql
@@ -12,7 +12,7 @@ CREATE TABLE IF NOT EXISTS Paths(
 	LastScrapedTimestamp BIGINT,
 	LastSuccessfulTimestamp BIGINT,
 	Path text NOT NULL,
-	BaseUrlID int NOT NULL,
+	BaseUrlID BIGINT NOT NULL,
 	PRIMARY KEY (PathID),
 	CONSTRAINT FK_BaseUrl FOREIGN KEY (BaseUrlID)
 	REFERENCES BaseURL(BaseUrlID)
@@ -26,7 +26,7 @@ CREATE TABLE IF NOT EXISTS Content(
 	ScrapeTimestamp BIGINT,
 	Content TEXT,
 	ContentType TEXT,
-	PathID int NOT NULL,
+	PathID BIGINT NOT NULL,
 	PRIMARY KEY (ContentID),
 	CONSTRAINT FK_Path FOREIGN KEY (PathID)
 	REFERENCES Paths(PathID)
@@ -36,8 +36,8 @@ CREATE TABLE IF NOT EXISTS Content(
 
 CREATE TABLE IF NOT EXISTS Links(
 	LinkId bigserial NOT NULL,
-	SourceContentId bigserial NOT NULL,
-	DestinationContentId bigserial NOT NULL,
+	SourceContentId BIGINT NOT NULL,
+	DestinationContentId BIGINT NOT NULL,
 	PRIMARY KEY (LinkId),
 	CONSTRAINT FK_Source FOREIGN KEY (SourceContentId)
 	REFERENCES Content(ContentID)
